rewrite indicator_text to take whether we want the actual result as input, text diff doesn't quite work yet and I IndicatorResultDisplay might be busted
